Refactor LogoAnimation and Layout components
This commit refactors the LogoAnimation and Layout components in the Svelte
app. The LogoAnimation component now uses the onMount hook to add the
handleScroll function that changes its border color when scrolled. The Layout
component removed the bind:this attribute in LogoAnimation component and
applied it directly to its class. Additionally, the manifest.json file's
theme_color property has been updated to match the app's new color.